District proximity- Expand nodes which has >0 selections on load Save Offline dates - request body removal
git-svn-id: https://svn.penske.com/svn/ecomm/webprojects/suppliermgmt/suppliermgmt/branches/dev@77253 1dbbddac-570e-0410-aece-e783461d2649

+++ JavaSource/com/penske/apps/buildmatrix/service/DefaultBuildMatrixSmcService.java
@@ -172,11 +172,11 @@ public class DefaultBuildMatrixSmcService implements BuildMatrixSmcService {
 	@Transactional
 	@Override
 	public void saveOfflineDates(BuildMatrixBodyPlant plantData) {
-		if(!plantData.getOfflineDateToRemove().isEmpty()) {
+		if(null!=plantData.getOfflineDateToRemove() && !plantData.getOfflineDateToRemove().isEmpty()) {
 			buildMatrixSmcDAO.removeOfflineDates(plantData.getOfflineDateToRemove());
 		}
 		
-		if(!plantData.getOfflineDates().isEmpty())
+		if(null != plantData.getOfflineDates() && !plantData.getOfflineDates().isEmpty())
 			buildMatrixSmcDAO.saveOfflineDates(plantData);
 	}
